backend/pyproject.toml
[project]
name = "rag-pipeline-lab-backend"
version = "0.1.0"
description = "Local-first RAG experimentation backend with FastAPI, SQLite, FAISS, and Ollama."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.8.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "numpy>=2.0.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "pypdf>=4.3.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
